+++ mRemoteV1/Connection/Protocol/IntegratedProgram.cs
@@ -5,6 +5,7 @@ using System.Diagnostics;
 using System.Drawing;
 using System.Threading;
 using System.Windows.Forms;
+using mRemoteNG.Messages;
 
 
 namespace mRemoteNG.Connection.Protocol
@@ -20,41 +21,52 @@ namespace mRemoteNG.Connection.Protocol
         #region Public Methods
 		public override bool Initialize()
 		{
-			if (InterfaceControl.Info != null)
-			{
-				_externalTool = Runtime.GetExtAppByName(Convert.ToString(InterfaceControl.Info.ExtApp));
-				_externalTool.ConnectionInfo = InterfaceControl.Info;
-			}
-					
-			return base.Initialize();
+		    if (InterfaceControl.Info == null) return base.Initialize();
+
+		    _externalTool = Runtime.GetExtAppByName(InterfaceControl.Info.ExtApp);
+		    _externalTool.ConnectionInfo = InterfaceControl.Info;
+
+		    return base.Initialize();
 		}
 				
 		public override bool Connect()
 		{
 			try
 			{
-				if (_externalTool.TryIntegrate == false)
+                Runtime.MessageCollector.AddMessage(MessageClass.InformationMsg, $"Attempting to start: {_externalTool.DisplayName}", true);
+
+                if (_externalTool.TryIntegrate == false)
 				{
 					_externalTool.Start(InterfaceControl.Info);
-					Close();
-					return false;
+                    /* Don't call close here... There's nothing for the override to do in this case since 
+                     * _process is not created in this scenario. When returning false, ProtocolBase.Close()
+                     * will be called - which is just going to call IntegratedProgram.Close() again anyway...
+                     * Close();
+                     */
+                    Runtime.MessageCollector.AddMessage(MessageClass.InformationMsg, $"Assuming no other errors/exceptions occurred immediately before this message regarding {_externalTool.DisplayName}, the next \"closed by user\" message can be ignored", true);
+                    return false;
 				}
 
-                ExternalToolArgumentParser argParser = new ExternalToolArgumentParser(_externalTool.ConnectionInfo);
-				_process = new Process();
-						
-				_process.StartInfo.UseShellExecute = true;
-                _process.StartInfo.FileName = argParser.ParseArguments(_externalTool.FileName);
-                _process.StartInfo.Arguments = argParser.ParseArguments(_externalTool.Arguments);
-						
-				_process.EnableRaisingEvents = true;
-				_process.Exited += ProcessExited;
+                var argParser = new ExternalToolArgumentParser(_externalTool.ConnectionInfo);
+			    _process = new Process
+			    {
+			        StartInfo =
+			        {
+			            UseShellExecute = true,
+			            FileName = argParser.ParseArguments(_externalTool.FileName),
+			            Arguments = argParser.ParseArguments(_externalTool.Arguments)
+			        },
+			        EnableRaisingEvents = true
+			    };
+
+
+			    _process.Exited += ProcessExited;
 						
 				_process.Start();
-				_process.WaitForInputIdle(Convert.ToInt32(Settings.Default.MaxPuttyWaitTime * 1000));
+				_process.WaitForInputIdle(Settings.Default.MaxPuttyWaitTime * 1000);
 						
-				int startTicks = Environment.TickCount;
-				while (_handle.ToInt32() == 0 & Environment.TickCount < startTicks + (Settings.Default.MaxPuttyWaitTime * 1000))
+				var startTicks = Environment.TickCount;
+				while (_handle.ToInt32() == 0 & Environment.TickCount < startTicks + Settings.Default.MaxPuttyWaitTime * 1000)
 				{
 					_process.Refresh();
 					if (_process.MainWindowTitle != "Default IME")
@@ -68,10 +80,10 @@ namespace mRemoteNG.Connection.Protocol
 				}
 						
 				NativeMethods.SetParent(_handle, InterfaceControl.Handle);
-				Runtime.MessageCollector.AddMessage(Messages.MessageClass.InformationMsg, Language.strIntAppStuff, true);
-				Runtime.MessageCollector.AddMessage(Messages.MessageClass.InformationMsg, string.Format(Language.strIntAppHandle, _handle.ToString()), true);
-				Runtime.MessageCollector.AddMessage(Messages.MessageClass.InformationMsg, string.Format(Language.strIntAppTitle, _process.MainWindowTitle), true);
-				Runtime.MessageCollector.AddMessage(Messages.MessageClass.InformationMsg, string.Format(Language.strIntAppParentHandle, InterfaceControl.Parent.Handle.ToString()), true);
+				Runtime.MessageCollector.AddMessage(MessageClass.InformationMsg, Language.strIntAppStuff, true);
+				Runtime.MessageCollector.AddMessage(MessageClass.InformationMsg, string.Format(Language.strIntAppHandle, _handle), true);
+				Runtime.MessageCollector.AddMessage(MessageClass.InformationMsg, string.Format(Language.strIntAppTitle, _process.MainWindowTitle), true);
+				Runtime.MessageCollector.AddMessage(MessageClass.InformationMsg, string.Format(Language.strIntAppParentHandle, InterfaceControl.Parent.Handle), true);
 						
 				Resize(this, new EventArgs());
 				base.Connect();
@@ -88,15 +100,12 @@ namespace mRemoteNG.Connection.Protocol
 		{
 			try
 			{
-				if (ConnectionWindow.InTabDrag)
-				{
-					return ;
-				}
+				if (ConnectionWindow.InTabDrag) return;
 				NativeMethods.SetForegroundWindow(_handle);
 			}
 			catch (Exception ex)
 			{
-				Runtime.MessageCollector.AddExceptionMessage(message: Language.strIntAppFocusFailed, ex: ex, logOnly: true);
+				Runtime.MessageCollector.AddExceptionMessage(Language.strIntAppFocusFailed, ex);
 			}
 		}
 				
@@ -104,45 +113,49 @@ namespace mRemoteNG.Connection.Protocol
 		{
 			try
 			{
-				if (InterfaceControl.Size == Size.Empty)
-				{
-					return ;
-				}
-                NativeMethods.MoveWindow(_handle, Convert.ToInt32(-SystemInformation.FrameBorderSize.Width), Convert.ToInt32(-(SystemInformation.CaptionHeight + SystemInformation.FrameBorderSize.Height)), InterfaceControl.Width + (SystemInformation.FrameBorderSize.Width * 2), InterfaceControl.Height + SystemInformation.CaptionHeight + (SystemInformation.FrameBorderSize.Height * 2), true);
+				if (InterfaceControl.Size == Size.Empty) return;
+                NativeMethods.MoveWindow(_handle, -SystemInformation.FrameBorderSize.Width, -(SystemInformation.CaptionHeight + SystemInformation.FrameBorderSize.Height), InterfaceControl.Width + SystemInformation.FrameBorderSize.Width * 2, InterfaceControl.Height + SystemInformation.CaptionHeight + SystemInformation.FrameBorderSize.Height * 2, true);
 			}
 			catch (Exception ex)
 			{
-				Runtime.MessageCollector.AddExceptionMessage(message: Language.strIntAppResizeFailed, ex: ex, logOnly: true);
+				Runtime.MessageCollector.AddExceptionMessage(Language.strIntAppResizeFailed, ex);
 			}
 		}
 				
 		public override void Close()
 		{
-			try
-			{
-				if (!_process.HasExited)
-				{
-					_process.Kill();
-				}
-			}
-			catch (Exception ex)
-			{
-				Runtime.MessageCollector.AddExceptionMessage(message: Language.strIntAppKillFailed, ex: ex, logOnly: true);
-			}
-					
-			try
-			{
-				if (!_process.HasExited)
-				{
-					_process.Dispose();
-				}
-			}
-			catch (Exception ex)
-			{
-				Runtime.MessageCollector.AddExceptionMessage(message: Language.strIntAppDisposeFailed, ex: ex, logOnly: true);
-			}
-					
-			base.Close();
+            /* only attempt this if we have a valid process object
+             * Non-integated tools will still call base.Close() and don't have a valid process object.
+             * See Connect() above... This just muddies up the log.
+             */
+            if (_process != null)
+		    {
+		        try
+		        {
+		            if (!_process.HasExited)
+		            {
+		                _process.Kill();
+		            }
+		        }
+		        catch (Exception ex)
+		        {
+		            Runtime.MessageCollector.AddExceptionMessage(Language.strIntAppKillFailed, ex);
+		        }
+
+		        try
+		        {
+		            if (!_process.HasExited)
+		            {
+		                _process.Dispose();
+		            }
+		        }
+		        catch (Exception ex)
+		        {
+		            Runtime.MessageCollector.AddExceptionMessage(Language.strIntAppDisposeFailed, ex);
+		        }
+		    }
+
+		    base.Close();
 		}
         #endregion
